+++ openmmmcmc/timing.py
