+++ tests/test_mcmc.py
@@ -8,7 +8,7 @@ from openmmtools import testsystems
 
 from pymbar import timeseries
 
-from openmm_mcmc.mcmc import HMCMove, GHMCMove, LangevinDynamicsMove, MonteCarloBarostatMove
+from openmmmcmc.mcmc import HMCMove, GHMCMove, LangevinDynamicsMove, MonteCarloBarostatMove
 import logging
 
 # Test various combinations of systems and MCMC schemes
@@ -25,7 +25,7 @@ debug = False # set to True only for manual debugging of this nose test
 
 def test_doctest():
     import doctest
-    from openmm_mcmc import mcmc
+    from openmmmcmc import mcmc
     doctest.testmod(mcmc)
 
 def test_minimizer_all_testsystems():
@@ -37,7 +37,7 @@ def test_minimizer_all_testsystems():
         
         testsystem = testsystem_class()
 
-        from openmm_mcmc import mcmc
+        from openmmmcmc import mcmc
         sampler_state = mcmc.SamplerState(testsystem.system, testsystem.positions)
 
         # Check if NaN.
@@ -82,15 +82,15 @@ def subtest_mcmc_expectation(testsystem, move_set):
     ndof = 3*system.getNumParticles() - system.getNumConstraints()
 
     # Create thermodynamic state
-    from openmm_mcmc.thermodynamics import ThermodynamicState
+    from openmmmcmc.thermodynamics import ThermodynamicState
     thermodynamic_state = ThermodynamicState(system=testsystem.system, temperature=temperature, pressure=pressure)
 
     # Create MCMC sampler.
-    from openmm_mcmc.mcmc import MCMCSampler
+    from openmmmcmc.mcmc import MCMCSampler
     sampler = MCMCSampler(thermodynamic_state, move_set=move_set, platform=platform)
 
     # Create sampler state.
-    from openmm_mcmc.mcmc import SamplerState
+    from openmmmcmc.mcmc import SamplerState
     sampler_state = SamplerState(system=testsystem.system, positions=testsystem.positions, platform=platform)
 
     # Equilibrate
